achievement model json->class conversion improved as to handle edge cases of null lists being returned or ints where doubles are supposed to be

+++ lib/data/profile/models/achievement_tile_model.dart
@@ -21,7 +21,7 @@ class AchievementTileModel extends AchievementTileEntity {
       title: json['title'] as String,
       description: json['description'] as String,
       quantity: json['quantity'] as int,
-      aspectRatio: json['aspect_ratio'] as double,
+      aspectRatio: json['aspect_ratio'].toDouble() as double,
     );
   }
 }
